# src/main/resources/MORGANMAZER-dragon-1.0-swagger.yaml
openapi: 3.0.0 
info: 
  version: '1.0' 
  title: 'Dragon Movie' 
  description: 'Dragon Movie project' 
servers: 
  - description: SwaggerHub API Auto Mocking 
    url: https://app.swaggerhub.com/apis/MORGANMAZER/dragon/1.0/
paths: 
  /movies: 
    post:
      summary: "Create a movie"
      description: "Allows an admin to create a movie" 
      parameters: 
        - name: access_token 
          in: header 
          description: Admin's access token for authorization. 
          required: true 
          schema: 
            type: string
      requestBody:
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Movie' 
        required: true
      responses:
        '201':
          description: "successfully created movie"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '403':
          description: "Forbidden. Only admins are allowed to perform this action."
    get: 
      summary: "Retrieve details about a movie"
      description: "Returns details about a specific movie" 
      parameters: 
        - name: title 
          in: query 
          description: the title of the movie to get details about 
          required: true 
          schema: 
            type: string 
            example: "Skyfall" 
      responses: 
        '200': 
          description: "successfully retrieved movie details" 
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Movie'
    delete: 
      summary: "Delete a movie"
      description: "Allows an admin to delete a movie" 
      parameters: 
        - name: access_token 
          in: header 
          description: Admin's access token for authorization. 
          required: true 
          schema: 
            type: string
        - name: title 
          in: query 
          description: the title of the movie to delete 
          required: true 
          schema: 
            type: string 
            example: "Skyfall"
      responses: 
        '200': 
          description: "successfully deleted movie" 
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/Movie'
        '403':
          description: "Forbidden. Only admins are allowed to perform this action."
    put: 
      summary: "Update a movie"
      description: "Allows an admin to update a movie" 
      parameters:
        - name: access_token 
          in: header 
          description: Admin's access token for authorization. 
          required: true 
          schema: 
            type: string
        - name: title 
          in: query 
          description: the title of the movie to update 
          required: true 
          schema: 
            type: string 
            example: "Skyfall" 
        - name: director 
          in: query 
          description: the updated director 
          required: false 
          schema: 
            type: string 
            example: "Sam Mendes" 
        - name: genre 
          in: query 
          description: the updated genre 
          required: false 
          schema: 
            type: string 
            example: "Action" 
        - name: rating 
          in: query 
          description: the updated rating 
          required: false 
          schema: 
            type: string 
            example: "PG-13" 
        - name: length 
          in: query
          description: the updated length 
          required: false 
          schema: 
            type: string 
            example: "2h23m" 
        - name: reviewScore 
          in: query 
          description: the updated review score 
          required: false 
          schema: 
            type: number 
            example: 7.8 
        - name: releaseDate 
          in: query 
          description: the updated release date 
          required: false 
          schema: 
            type: string 
            example: "2012-10-26"
        - name: currentlyPlaying 
          in: query 
          description: the updated value for if the movie is currently playing 
          required: false 
          schema: 
            type: boolean 
            example: false 
        - name: upcomingRelease 
          in: query 
          description: the updated value for if the move is an upcoming release 
          required: false 
          schema: 
            type: boolean 
            example: false
      responses:
        '200': 
          description: "successfully updated movie" 
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Movie' 
        '403':
          description: "Forbidden. Only admins are allowed to perform this action."
...
components: 
  schemas: 
...
    Movie:
      type: object 
      properties: 
        title: 
          type: string 
          example: "Oppenheimer" 
        director: 
          type: string 
          example: "Christopher Nolan" 
        genre: 
          type: string 
          example: "Thriller" 
        rating: 
          type: string 
          example: "R" 
        length: 
          type: string 
          example: "3h5m" 
        releaseDate: 
          type: string 
          format: date 
          example: 2023-07-21 
        reviewScore: 
          type: number 
          example: 8.9 
        currentlyPlaying: 
          type: boolean 
          example: true 
        upcomingRelease: 
          type: boolean 
          example: false 
...